Lil more cleanup
