+++ src/audio.rs
@@ -1,0 +1,22 @@
+use rodio::cpal::traits::{DeviceTrait, HostTrait};
+use rodio::OutputStream;
+use rodio::*;
+
+pub fn get_output_stream(
+    device_name: &str,
+) -> (OutputStream, OutputStreamHandle) {
+    let host = cpal::default_host();
+    let devices = host.output_devices().unwrap();
+
+    let (mut _stream, mut stream_handle) = OutputStream::try_default().unwrap();
+    for device in devices {
+        let dev: rodio::Device = device.into();
+        let dev_name: String = dev.name().unwrap();
+        if dev_name == device_name {
+            println!("Device found: {}", dev_name);
+            (_stream, stream_handle) =
+                OutputStream::try_from_device(&dev).unwrap();
+        }
+    }
+    return (_stream, stream_handle);
+}
